refactor(tools): move web tools from legacy/tools/ to agent/tools/
createWebFetchTool / createWebSearchTool are AgentTool factories used
only by createAgentTools. Belong with the rest of the agent tools.

- mv legacy/tools/web.{ts,test.ts} → agent/tools/web.{ts,test.ts}
- update 4 import sites (agent/tools/index.ts, its test, e2e-smoke,
  legacy/tools/index.ts barrel)
- drop stale src/tools/web.ts file-header comment

No behavior change.
